pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inventario-faces"
dynamic = ["version"]
description = "Aplicativo desktop forense para inventario assistido de faces em imagens e videos."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
  { name = "Codex" }
]
dependencies = [
  "numpy>=1.26",
  "opencv-python>=4.10",
  "PySide6>=6.8",
  "PyYAML>=6.0",
  "onnxruntime>=1.22",
  "python-docx>=1.2"
]

[project.optional-dependencies]
face = [
  "insightface>=0.7.3; python_version < '3.13'"
]
build = [
  "pyinstaller>=6.11"
]

[project.scripts]
inventario-faces = "inventario_faces.__main__:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.dynamic]
version = { attr = "inventario_faces.__version__" }

[tool.setuptools.package-data]
inventario_faces = [
  "reporting/templates/*.tex",
  "config/*.yaml"
]
